training duration

+++ cpp/core/sparse_mat.hpp
@@ -94,7 +94,7 @@ public:
       uint64_t checking_index = rank * proc_col_width;
       //      uint64_t checking_index = 1;
       uint64_t checking_end_index =
-          std::min(((rank + 1) * proc_col_width) - 1, gCols - 1);
+          std::min(static_cast<uint64_t>((rank + 1) * proc_col_width) - 1, gCols - 1);
 
       auto startIt = std::find_if(coords.begin(), coords.end(),
                                   [&checking_index](const auto &tuple) {
